fix: move key handlers to useHandleKeyDown hook
