+++ src/components/Calendar/doubleCalendar/hooks/index.js
@@ -4,3 +4,4 @@ export { default as useFormattedMonth } from './useFormattedMonth';
 export { default as useMoveFocusedDay } from './useMoveFocusedDay';
 export { default as useMoveFocusedMonth } from './useMoveFocusedMonth';
 export { default as useDisabledControls } from './useDisabledControls';
+export { default as useHandleKeyDown } from './useHandleKeyDown';
