+++ src/components/Calendar/doubleCalendar/index.js
@@ -1,18 +1,6 @@
-import React, { useMemo, useState, useCallback, useEffect } from 'react';
+import React, { useState, useCallback, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useUniqueIdentifier } from '../../../libs/hooks';
-import {
-    UP_KEY,
-    DOWN_KEY,
-    RIGHT_KEY,
-    LEFT_KEY,
-    HOME_KEY,
-    END_KEY,
-    PAGEUP_KEY,
-    PAGEDN_KEY,
-    SPACE_KEY,
-    ENTER_KEY,
-} from '../../../libs/constants';
 import { Provider } from '../context';
 import RightIcon from '../icons/rightArrow';
 import LeftIcon from '../icons/leftArrow';
@@ -23,17 +11,10 @@ import {
     getFirstDayMonth,
     addMonths,
     getNextFocusedDate,
-    isSameMonth,
     isEmptyRange,
     isDateBelowLimit,
 } from '../helpers';
-import {
-    useYearsRange,
-    useMoveFocusedDay,
-    useFormattedMonth,
-    useMoveFocusedMonth,
-    useDisabledControls,
-} from './hooks';
+import { useYearsRange, useDisabledControls, useFormattedMonth, useHandleKeyDown } from './hooks';
 import MonthHeader from './monthHeader';
 import StyledControlsContainer from '../styled/controlsContainer';
 import StyledArrowButton from '../styled/arrowButton';
@@ -74,9 +55,17 @@ export default function DoubleCalendar(props) {
         currentMonth,
         rightCalendarMonth,
     );
-    const moveFocusedDay = useMoveFocusedDay(focusedDate, currentMonth, minDate, maxDate);
-    const moveFocusedMonth = useMoveFocusedMonth(focusedDate, minDate, maxDate);
-
+    const handleKeyDown = useHandleKeyDown(
+        focusedDate,
+        currentMonth,
+        rightCalendarMonth,
+        minDate,
+        maxDate,
+        onChange,
+        enableNavKeys,
+        setFocusedDate,
+        setCurrentMonth,
+    );
     const handleOnDayFocus = () => setEnableNavKeys(true);
     const handleOnDayBlur = () => setEnableNavKeys(false);
     const handleOnDayHover = useCallback(
@@ -129,105 +118,6 @@ export default function DoubleCalendar(props) {
             setCurrentMonth(addMonths(newMonth, -1));
         },
         [rightCalendarMonth, value],
-    );
-
-    const keyHandlerMap = useMemo(
-        () => ({
-            [UP_KEY]: () => {
-                const result = moveFocusedDay(-7);
-                setFocusedDate(result.day);
-            },
-            [DOWN_KEY]: () => {
-                const result = moveFocusedDay(7);
-                setFocusedDate(result.day);
-            },
-            [LEFT_KEY]: () => {
-                const result = moveFocusedDay(-1);
-                setFocusedDate(result.day);
-            },
-            [RIGHT_KEY]: () => {
-                const result = moveFocusedDay(1);
-                setFocusedDate(result.day);
-            },
-            [HOME_KEY]: () => {
-                const result = moveFocusedDay(-focusedDate.getDay());
-                setFocusedDate(result.day);
-            },
-            [END_KEY]: () => {
-                const result = moveFocusedDay(6 - focusedDate.getDay());
-                setFocusedDate(result.day);
-            },
-            [PAGEUP_KEY]: () => {
-                const result = moveFocusedMonth(-1);
-                setFocusedDate(result.day);
-                if (
-                    !isSameMonth(result.month, rightCalendarMonth) &&
-                    !isSameMonth(result.month, currentMonth)
-                ) {
-                    setCurrentMonth(result.month);
-                }
-            },
-            [PAGEDN_KEY]: () => {
-                const result = moveFocusedMonth(1);
-                setFocusedDate(result.day);
-                if (
-                    !isSameMonth(result.month, rightCalendarMonth) &&
-                    !isSameMonth(result.month, currentMonth)
-                ) {
-                    setCurrentMonth(rightCalendarMonth);
-                }
-            },
-            [SPACE_KEY]: () => onChange(new Date(focusedDate)),
-            [ENTER_KEY]: () => onChange(new Date(focusedDate)),
-        }),
-        [currentMonth, focusedDate, moveFocusedDay, moveFocusedMonth, onChange, rightCalendarMonth],
-    );
-
-    const keyHandlerMapAlt = useMemo(
-        () => ({
-            [HOME_KEY]: () => {
-                const result = moveFocusedDay(-focusedDate.getDay());
-                setFocusedDate(result.day);
-            },
-            [END_KEY]: () => {
-                const result = moveFocusedDay(6 - focusedDate.getDay());
-                setFocusedDate(result.day);
-            },
-            [PAGEUP_KEY]: () => {
-                const result = moveFocusedMonth(-12);
-                setFocusedDate(result.day);
-                if (isSameMonth(result.month, rightCalendarMonth)) {
-                    setCurrentMonth(addMonths(result.month, -1));
-                } else {
-                    setCurrentMonth(result.month);
-                }
-            },
-            [PAGEDN_KEY]: () => {
-                const result = moveFocusedMonth(12);
-                setFocusedDate(result.day);
-                if (isSameMonth(result.month, rightCalendarMonth)) {
-                    setCurrentMonth(addMonths(result.month, -1));
-                } else {
-                    setCurrentMonth(result.month);
-                }
-            },
-        }),
-        [focusedDate, moveFocusedDay, moveFocusedMonth, rightCalendarMonth],
-    );
-
-    const handleKeyDown = useCallback(
-        event => {
-            if (enableNavKeys) {
-                const { keyCode, altKey } = event;
-                const keyHandler = altKey ? keyHandlerMapAlt : keyHandlerMap;
-                if (keyHandler[keyCode]) {
-                    event.preventDefault();
-                    event.stopPropagation();
-                    keyHandler[keyCode]();
-                }
-            }
-        },
-        [enableNavKeys, keyHandlerMap, keyHandlerMapAlt],
     );
 
     useEffect(() => {
